requirements.txt
Flask
#Tenacity
bs4
python-dateutil
